+++ econtext/src/test/java/com/adyen/checkout/econtext/TestEContextComponent.kt
@@ -11,6 +11,7 @@ package com.adyen.checkout.econtext
 import com.adyen.checkout.action.internal.DefaultActionHandlingComponent
 import com.adyen.checkout.action.internal.ui.GenericActionDelegate
 import com.adyen.checkout.components.core.internal.ComponentEventHandler
+import com.adyen.checkout.econtext.internal.EContextComponent
 import com.adyen.checkout.econtext.internal.ui.EContextDelegate
 
 internal class TestEContextComponent internal constructor(
